Guard against boolean return value from get_latest_events()
We've observed an issue in a test environment where the `Alert_Manager::get_latest_events` method returned `false` (indicating that connectivity was lost), which is a TypeError on recent versions of PHP because `count` requires its argument to be countable. When using `declare( strict_types=1 );`, this error prevents the admin from loading when this situation occurs.

This patch ensures that the admin will load even if the get_latest_events method returns false.